refactor: move docs sectioncontent and content model into @repo/shared (#1568)
