+++ app/components/Docs/components/SectionHero/components/HeroImage/heroImage.styles.ts
@@ -1,6 +1,6 @@
-import { PADDING_Y as SECTION_PADDING_Y } from "@/components/Docs/components/SectionContent/sectionContent.styles";
 import { PADDING_Y_BOTTOM as HERO_SECTION_PADDING_Y_BOTTOM } from "@/components/Docs/components/SectionHero/sectionHero.styles";
 import styled from "@emotion/styled";
+import { PADDING_Y as SECTION_PADDING_Y } from "@repo/shared/views/docs/components/SectionContent/sectionContent.styles";
 
 const PADDING_Y = 56;
