+++ app/components/Docs/components/SectionHero/sectionHero.styles.ts
@@ -1,7 +1,3 @@
-import {
-  CONTENT_TYPE,
-  FrontmatterProps,
-} from "@/docs/common/frontmatter/types";
 import { css } from "@emotion/react";
 import styled from "@emotion/styled";
 import { SectionHero } from "@repo/shared/components/layout/SectionHero/sectionHero";
@@ -9,6 +5,10 @@ import {
   Head,
   SectionLayout,
 } from "@repo/shared/components/layout/SectionHero/sectionHero.styles";
+import {
+  CONTENT_TYPE,
+  FrontmatterProps,
+} from "@repo/shared/views/docs/common/frontmatter/types";
 
 export const PADDING_Y_BOTTOM = 90;
