+++ app/components/Docs/components/SectionOverview/sectionOverview.tsx
@@ -2,9 +2,9 @@ import { CardActionArea } from "@databiosphere/findable-ui/lib/components/common
 import { CardTitle } from "@databiosphere/findable-ui/lib/components/common/Card/components/CardTitle/cardTitle";
 import { TYPOGRAPHY_PROPS } from "@databiosphere/findable-ui/lib/styles/common/mui/typography";
 import { CardMedia, Stack, Typography } from "@mui/material";
+import { Props } from "@repo/shared/views/docs/components/SectionOverview/types";
 import { JSX } from "react";
 import { StyledCard, StyledStack } from "./sectionOverview.styles";
-import { Props } from "./types";
 
 export const SectionOverview = ({ overview }: Props): JSX.Element | null => {
   if (!overview) return null;
