+++ app/docs/common/staticGeneration/utils.ts
@@ -1,5 +1,5 @@
+import { StaticProps } from "@repo/shared/views/docs/common/staticGeneration/types";
 import { GetStaticPropsResult } from "next/types";
-import { StaticProps } from "./types";
 
 /**
  * Processes the outline for the given static props.
