+++ app/views/AboutView/aboutView.tsx
@@ -1,7 +1,7 @@
 import { SectionContentCard } from "@/components/common/Card/components/SectionContentCard/sectionContentCard";
-import { SectionContent } from "@/components/Docs/components/SectionContent/sectionContent";
 import { StyledStack } from "@/components/Docs/components/SectionContentCards/sectionContentCards.styles";
 import { StyledSectionHero } from "@/components/Docs/components/SectionHero/sectionHero.styles";
+import { SectionContent } from "@repo/shared/views/docs/components/SectionContent/sectionContent";
 import { ComponentProps, Fragment, JSX } from "react";
 
 interface AboutViewProps {
