+++ packages/shared/views/docs/components/SectionContent/sectionContent.tsx
@@ -1,8 +1,8 @@
-import { Section } from "@/components/content/content.styles";
-import { StaticProps } from "@/docs/common/staticGeneration/types";
 import { PANEL_BACKGROUND_COLOR } from "@databiosphere/findable-ui/lib/components/Layout/components/ContentLayout/common/entities";
 import { ContentGrid } from "@databiosphere/findable-ui/lib/components/Layout/components/ContentLayout/contentLayout.styles";
-import { ContentViewProps } from "@databiosphere/findable-ui/lib/views/ContentView/contentView";
+import type { ContentViewProps } from "@databiosphere/findable-ui/lib/views/ContentView/contentView";
+import type { StaticProps } from "@repo/shared/views/docs/common/staticGeneration/types";
+import { Section } from "@repo/shared/views/docs/content.styles";
 import { JSX } from "react";
 import {
   StyledContent,
