+++ packages/shared/views/docs/components/SectionOverview/types.ts
@@ -1,4 +1,4 @@
-import { ComponentProps } from "react";
+import type { ComponentProps } from "react";
 
 export interface Overview {
   date: string;
